Allow (optional) choice of scoped selectors
Sorry for the unprofessional PR, just edited on github directly and didn't test it, but I think it should work.

The idea is that we could give the `scopedCSS` property of the options object for `riot.compile()` a value other than `true` or `false` (those stay default and are ok), e.g. a string either `'tag'` or `'is'` which will then in turn only scope the CSS selectors by the chosen method.

This allows us to reduce CSS size by cutting down on the bloat the CSS scoping does by adding both the `tag` and the `[is="<tag>"]` prefixes to each selector within a components CSS. I believe (and certainly do personally) people probably use the `is` approach or the `tag` approach, but not mix both and if they do, the defaults are still fine adding both prefixes to selectors. For all others who use one approach consistently (I'm in the `is` camp) this change could give some nice size savings in the generated CSS.

Maybe you see a better way of doing this as well?

Thanks for your great work on riot and for looking into this!